% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norTEPtest.R
\name{C_matrix_MC_par}
\alias{C_matrix_MC_par}
\title{Monte Carlo approximation of C matrix}
\usage{
C_matrix_MC_par(
  l,
  N_per_core = 1e+08,
  block_size = 1e+05,
  n_cores = detectCores() - 1,
  seed = 42
)
}
\arguments{
\item{l}{Integer, maximum Hermite degree}

\item{N_per_core}{Number of samples per core}

\item{block_size}{Size of each block}

\item{n_cores}{Number of cores used}

\item{seed}{Random seed}
}
\value{
Symmetric matrix C of dimension (l+1) x (l+1)
}
\description{
Computes the covariance matrix C by parallel Monte Carlo simulation
}
