% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norTEPtest.R
\name{eval_hvec_stable}
\alias{eval_hvec_stable}
\title{Normalized Hermite polynomials}
\usage{
eval_hvec_stable(x_vec, l)
}
\arguments{
\item{x_vec}{Numeric vector}

\item{l}{Integer}
}
\value{
A matrix with rows corresponding to degrees 0..l
}
\description{
Computes the normalized Hermite polynomials
of degrees 0 to l at the entries of the vector x_vec
}
