% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norTEPtest.R
\name{TEPest}
\alias{TEPest}
\title{TEP-based CvM statistic}
\usage{
TEPest(X, type = "skew", l = 10)
}
\arguments{
\item{X}{Numeric vector, the sample}

\item{type}{Character specifying the alternative of focusing}

\item{l}{Integer, number of Hermite polynomials evaluated at the sample points -1}
}
\value{
Numeric value of the test statistic
}
\description{
Computes the TEP-based Cramér–von Mises statistic
}
