% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norTEPtest.R
\name{TEPnor}
\alias{TEPnor}
\title{Normality tests based on TEP's}
\usage{
TEPnor(X, type = "skew", l = 10, nrep = 1000)
}
\arguments{
\item{X}{Vector containing the elements of the sample}

\item{type}{Character specifying the alternative of focusing:
\itemize{
\item \code{"skew"} — Test focused to detect skewness
\item \code{"kurt"} — Test focused to detect changes in kurtosis
}}

\item{l}{Integer, Number of Hermite polynomials evaluated at the sample points -1}

\item{nrep}{Integer, Number of Monte Carlo evaluations to estimate the p-value
(defalts to 1000)}
}
\value{
A list with
\itemize{
\item \code{"pval"} — The p-value of the sample estimated by MC
\item \code{"type"} — The alternative of focusing
\item \code{"est"} — The value of the test statistic
}
}
\description{
Performs tests of normality consistent
and focused to detect skewness (type='skew', the default)
or kurtosis (type='kurt')
}
\details{
The tests statistics are quadratic forms in the vector of evaluations
of the normalized Hermite polynomials at the scaled sample points,
introduced by the authors in
\emph{Tests of normality based on Transformed Empirical Processes},
Methodology and Computing in Applied Probability, Kluwer Academic
Publishers,vol.5, pp 309-335, 2003.
}
\examples{
\dontrun{
set.seed(42)
X <- rnorm(100)
TEPnor(X)
}

}
